Account for additional embedded control commands in structs
Currently when we parse structs, we allow for parameterless control commands
like `.word` or `.byte`. But it's also common to use slightly more complex
commands like `.tag` or `.res` that are followed by an identifier or a number.

This patch crudely allows for these general shapes by allowing optional
identifiers or numbers following a macro token type.